prepare stdin
