+++ dmsgpty/cli_windows.go
@@ -8,8 +8,6 @@ import (
 	"github.com/ActiveState/termtest/conpty"
 	"io"
 	"os"
-
-	"golang.org/x/sys/windows"
 )
 
 // ptyResizeLoop informs the remote of changes to the local CLI terminal window size.
@@ -18,9 +16,8 @@ func ptyResizeLoop(_ context.Context, ptyC *PtyClient) error {
 	return nil
 }
 
-// getPtySize obtains the size of the local terminal.
-func getPtySize(_ *os.File) (*windows.Coord, error) {
-	return getSize()
+func (cli *CLI) prepareStdin() (restore func(), err error) {
+	return conpty.InitTerminal(true)
 }
 
 // servePty serves a pty connection via the dmsgpty-host.
@@ -32,25 +29,19 @@ func (cli *CLI) servePty(ctx context.Context, ptyC *PtyClient, cmd string, args 
 		WithField("cmd", fmt.Sprint(append([]string{cmd}, args...))).
 		Infof("Executing...")
 
-	f, err := conpty.InitTerminal(false)
-	if err != nil {
-		return err
-	}
-	defer f()
-
 	if err := ptyC.Start(cmd, args...); err != nil {
 		return fmt.Errorf("failed to start command on pty: %v", err)
 	}
 
-	// Window resize loop.
-	go func() {
-		defer cancel()
-		if err := ptyResizeLoop(ctx, ptyC); err != nil {
-			cli.Log.
-				WithError(err).
-				Warn("Window resize loop closed with error.")
-		}
-	}()
+	// TODO: Window resize loop.
+	//go func() {
+	//	defer cancel()
+	//	if err := ptyResizeLoop(ctx, ptyC); err != nil {
+	//		cli.Log.
+	//			WithError(err).
+	//			Warn("Window resize loop closed with error.")
+	//	}
+	//}()
 
 	// Write loop.
 	go func() {
